+++ src/XmlDocMarkdown.Core/ExternalDocumentation.cs
@@ -1,0 +1,13 @@
+namespace XmlDocMarkdown.Core
+{
+	/// <summary>
+	/// Configures external documentation.
+	/// </summary>
+	public sealed class ExternalDocumentation
+	{
+		/// <summary>
+		/// The namespace to configure.
+		/// </summary>
+		public string Namespace { get; set; }
+	}
+}