+++ VersionHistory.md
@@ -7,10 +7,12 @@ is updated in preparation for publishing an updated NuGet package.
 
 Prefix the description of the change with `[major]`, `[minor]` or `[patch]` in accordance with [SemVer](http://semver.org).
 
-* [minor] Add `--settings` command-line option.
-* [minor] Add minimal external documentation support (same repository).
-
 ## Released
+
+### 1.3.0
+
+* Add `--settings` command-line option.
+* Add minimal external documentation support (same repository).
 
 ### 1.2.1